+++ src/settings/MappingProfiles/detailsSections/edit/OrderDetailSection/OrderInformation.js
@@ -39,6 +39,7 @@ import {
   getSubfieldName,
   onAdd,
   onRemove,
+  renderFieldLabelWithInfo,
 } from '../../utils';
 import {
   DEFAULT_PO_LINES_LIMIT_VALUE,
@@ -109,6 +110,11 @@ const OrderInformationComponent = ({
         setReferenceTables(ORDER_INFO_FIELDS_MAP.PO_LINES_LIMIT, `"${purchaseOrderLinesLimitValue}"`);
       });
   }, []); // eslint-disable-line react-hooks/exhaustive-deps
+
+  const purchaseOrderStatusLabel = renderFieldLabelWithInfo(
+    `${TRANSLATION_ID_PREFIX}.order.orderInformation.field.purchaseOrderStatus`,
+    `${TRANSLATION_ID_PREFIX}.order.orderInformation.field.purchaseOrderStatus.info`,
+  );
 
   const isApprovalRequiredValue = useMemo(
     () => {
@@ -200,7 +206,7 @@ const OrderInformationComponent = ({
           <AcceptedValuesField
             component={TextField}
             name={ORDER_INFO_FIELDS_MAP.PO_STATUS}
-            label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.order.orderInformation.field.purchaseOrderStatus`} />}
+            label={purchaseOrderStatusLabel}
             optionValue="value"
             optionLabel="label"
             wrapperLabel={`${TRANSLATION_ID_PREFIX}.wrapper.acceptedValues`}